Skip rule api: Move StationConfigDTO to StationDTO
