+++ pom.xml
@@ -24,6 +24,7 @@
 		<java.version>1.8</java.version>
 		<org.mapstruct.version>1.2.0.CR2</org.mapstruct.version>
 		<springfox-swagger.version>2.7.0</springfox-swagger.version>
+        <maven.test.skip>true</maven.test.skip>
 	</properties>
 
 	<dependencies>
