+++ src/main/java/com/mgmtp/radio/aop/StationAspect.java
@@ -17,24 +17,24 @@ import java.util.Map;
 @Component
 public class StationAspect {
 
-	MessageChannel registerChannel;
-
-	public StationAspect(MessageChannel registerChannel) {
-		this.registerChannel = registerChannel;
-	}
-
-	@AfterReturning(value = "execution(* com.mgmtp.radio.service.station.StationServiceImpl.checkAndSkipSongIfNeeded(..))", returning = "monoSongDTO")
-	public void updatePlaylist(Mono<SongDTO> monoSongDTO) {
-		monoSongDTO.map(m -> {
-			if(m.isSkipped()) {
-				Map<String, Object> updatePram = new HashMap<>();
-				updatePram.put(StationEventDataKeys.station_id.name(), m.getStationId());
-				updatePram.put(StationEventDataKeys.event_id.name(), SubscriptionEvents.updatePlaylist.name());
-
-				registerChannel.send(new GenericMessage<>(updatePram));
-			}
-			return m;
-		});
-
-	}
+//	MessageChannel registerChannel;
+//
+//	public StationAspect(MessageChannel registerChannel) {
+//		this.registerChannel = registerChannel;
+//	}
+//
+//	@AfterReturning(value = "execution(* com.mgmtp.radio.service.station.StationServiceImpl.checkAndSkipSongIfNeeded(..))", returning = "monoSongDTO")
+//	public void updatePlaylist(Mono<SongDTO> monoSongDTO) {
+//		monoSongDTO.map(m -> {
+//			if(m.isSkipped()) {
+//				Map<String, Object> updatePram = new HashMap<>();
+//				updatePram.put(StationEventDataKeys.station_id.name(), m.getStationId());
+//				updatePram.put(StationEventDataKeys.event_id.name(), SubscriptionEvents.updatePlaylist.name());
+//
+//				registerChannel.send(new GenericMessage<>(updatePram));
+//			}
+//			return m;
+//		});
+//
+//	}
 }
