+++ src/main/java/com/mgmtp/radio/dto/station/StationConfigurationDTO.java
@@ -1,5 +1,6 @@
 package com.mgmtp.radio.dto.station;
 
+import com.mgmtp.radio.dto.skipRule.SkipRuleDTO;
 import lombok.Data;
 import lombok.NoArgsConstructor;
