+++ src/main/java/com/mgmtp/radio/mapper/station/StationMapper.java
@@ -2,7 +2,6 @@ package com.mgmtp.radio.mapper.station;
 
 import com.mgmtp.radio.domain.station.Station;
 import com.mgmtp.radio.dto.station.StationDTO;
-import org.mapstruct.DecoratedWith;
 import org.mapstruct.Mapper;
 import org.mapstruct.Mapping;
 import org.mapstruct.factory.Mappers;
