+++ src/main/java/com/mgmtp/radio/mapper/stationConfiguration/StationConfigurationMapper.java
@@ -2,7 +2,7 @@ package com.mgmtp.radio.mapper.stationConfiguration;
 
 import com.mgmtp.radio.domain.station.SkipRule;
 import com.mgmtp.radio.domain.station.StationConfiguration;
-import com.mgmtp.radio.dto.station.SkipRuleDTO;
+import com.mgmtp.radio.dto.skipRule.SkipRuleDTO;
 import com.mgmtp.radio.dto.station.StationConfigurationDTO;
 import org.mapstruct.Mapper;
 import org.mapstruct.factory.Mappers;
