+++ src/main/java/com/mgmtp/radio/service/station/StationServiceImpl.java
@@ -1,10 +1,8 @@
 package com.mgmtp.radio.service.station;
 
 import com.mgmtp.radio.domain.station.SkipRule;
-import com.mgmtp.radio.domain.station.Song;
 import com.mgmtp.radio.domain.station.Station;
 import com.mgmtp.radio.domain.station.StationConfiguration;
-import com.mgmtp.radio.dto.station.SkipRuleDTO.InvalidRuleTypeDtoException;
 import com.mgmtp.radio.dto.station.SongDTO;
 import com.mgmtp.radio.dto.station.StationConfigurationDTO;
 import com.mgmtp.radio.dto.station.StationDTO;
@@ -72,11 +70,7 @@ public class StationServiceImpl implements StationService {
 				}
 			} else {
 				double downvotePercent = 0;
-				try {
-					downvotePercent = calcCurrentSongDislikePercent(songDTO, new StationDTO());
-				} catch (InvalidRuleTypeDtoException e) {
-					e.printStackTrace();
-				}
+				downvotePercent = calcCurrentSongDislikePercent(songDTO, new StationDTO());
 				if (downvotePercent > DOWN_VOTE_THRES_PERCENT) {
 					isSkipped = true;
 				}
@@ -107,8 +101,8 @@ public class StationServiceImpl implements StationService {
     }
 
 	public Flux<StationDTO> getAll() {
-        return stationRepository.findAll()
-                .map(stationMapper::stationToStationDTO);
+		    return stationRepository.findAll()
+			    .map(stationMapper::stationToStationDTO);
     }
 
     @Override
