+++ src/main/resources/application.properties
@@ -10,7 +10,7 @@ application.name=team-radio-services
 security.basic.enabled=false
 
 #mongo db
-spring.data.mongodb.uri=mongodb://radio_admin:VvV2geQ4Pp3XmpN5UM7rmeLmmZwATYVs@ds137090.mlab.com:37090/new_radio_db_dev
+spring.data.mongodb.uri=mongodb://radio_admin:VvV2geQ4Pp3XmpN5UM7rmeLmmZwATYVs@localhost:27017/new_radio_db_dev
 
 #upload files
 cloudinary.api.cloud.name=drvjxl2xr
