+++ src/test/java/com/mgmtp/radio/mapper/stationConfiguration/StationConfigurationMapperTest.java
@@ -1,7 +1,7 @@
 package com.mgmtp.radio.mapper.stationConfiguration;
 
 import com.mgmtp.radio.domain.station.SkipRule;
-import com.mgmtp.radio.dto.station.SkipRuleDTO;
+import com.mgmtp.radio.dto.skipRule.SkipRuleDTO;
 import org.junit.Test;
 
 import static org.junit.Assert.*;
