+++ src/test/java/com/mgmtp/radio/service/station/CreateStationImplTest.java
@@ -1,7 +1,6 @@
 package com.mgmtp.radio.service.station;
 
 import com.mgmtp.radio.domain.station.Station;
-import com.mgmtp.radio.dto.station.SkipRuleDTO.InvalidRuleTypeDtoException;
 import com.mgmtp.radio.dto.station.StationDTO;
 import com.mgmtp.radio.mapper.station.StationMapper;
 import com.mgmtp.radio.mapper.stationConfiguration.StationConfigurationMapper;
@@ -40,7 +39,7 @@ public class CreateStationImplTest {
     }
 
     @Test
-    public void createSuccess() throws InvalidRuleTypeDtoException {
+    public void createSuccess() {
         //given
         StationDTO stationDTO = new StationDTO();
         stationDTO.setId("station1");
