+++ src/test/java/com/mgmtp/radio/service/station/GetStationByIdImplTest.java
@@ -1,7 +1,6 @@
 package com.mgmtp.radio.service.station;
 
 import com.mgmtp.radio.domain.station.Station;
-import com.mgmtp.radio.dto.station.SkipRuleDTO.InvalidRuleTypeDtoException;
 import com.mgmtp.radio.dto.station.SongDTO;
 import com.mgmtp.radio.dto.station.StationDTO;
 import com.mgmtp.radio.mapper.station.StationMapper;
@@ -47,7 +46,7 @@ public class GetStationByIdImplTest {
     }
 
     @Test
-    public void GetStationByIdSuccess() throws InvalidRuleTypeDtoException {
+    public void GetStationByIdSuccess() {
         //given
         SongDTO songDTO = new SongDTO();
         songDTO.setId("song1");
